refactor: Drop `roles` package
