+++ pkg/nat/errors.go
