+++ pkg/packager/archive.go
