+++ pkg/packager/extract.go
