+++ pkg/peer/migrate_from.go
@@ -14,10 +14,10 @@ import (
 	"syscall"
 
 	"github.com/loopholelabs/drafter/internal/utils"
-	"github.com/loopholelabs/drafter/pkg/roles/mounter"
-	"github.com/loopholelabs/drafter/pkg/roles/registry"
-	"github.com/loopholelabs/drafter/pkg/roles/snapshotter"
-	"github.com/loopholelabs/drafter/pkg/roles/terminator"
+	"github.com/loopholelabs/drafter/pkg/mounter"
+	"github.com/loopholelabs/drafter/pkg/registry"
+	"github.com/loopholelabs/drafter/pkg/snapshotter"
+	"github.com/loopholelabs/drafter/pkg/terminator"
 	"github.com/loopholelabs/goroutine-manager/pkg/manager"
 	"github.com/loopholelabs/silo/pkg/storage"
 	"github.com/loopholelabs/silo/pkg/storage/config"
