+++ pkg/registry/errors.go
