+++ pkg/registry/events.go
