+++ pkg/registry/open.go
