+++ pkg/terminator/errors.go
